Fix AGS Legend display and improve JSON handling
Updated changelog with a fix for AGS Legend not displaying when JSON values are not strings. Incremented version to 7.25.2501 and added hover highlight feature for map responses. Removed redundant condition in `ClientIdAndSecretAuthenticationMiddleware.cs`. Enhanced `Legend.cs` with a new JSON converter for the `Values` property. Introduced `StringArrayConverter` for custom serialization and deserialization of string arrays.